[Bronze II] Title: 숫자, Time: 840 ms, Memory: 30084 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -4,8 +4,8 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        long a = sc.nextInt();
-        long b = sc.nextInt();
+        long a = sc.nextLong();
+        long b = sc.nextLong();
 
         if(a > b) {
             long temp = a;
